Succesful Debug build on x64

+++ Mini_Maniacs/Stream.cpp
@@ -52,7 +52,7 @@ glm::vec2 Stream::ReadVector(void)
 SDL_Color Stream::ReadColor(void) 
 {
   SDL_Color c = {};
-  glm::vec4 in = {};
+  glm::ivec4 in = {};
   fileStream >> in.r;
   fileStream >> in.g;
   fileStream >> in.b;
